+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fixed: Disable duress pin mistakenly disables pin-login for entire account.
+
 ## 2.27.3 (2025-05-12)
 
 - fixed: Bug causing duress account creation from `loginWithKey` while in duress mode.